Refactor classes to satisfy UML diagram
