+++ app/src/main/java/com/ss20/se2/monopoly/pojo/models/deeds/CommunityCard.java
@@ -1,16 +1,8 @@
 package com.ss20.se2.monopoly.pojo.models.deeds;
 
-public class CommunityCard extends Deed {
-
-    private String description;
+public class CommunityCard extends Card {
 
     public CommunityCard(String name, String description) {
-        super(name);
-
-        this.description = description;
-    }
-
-    public String getDescription() {
-        return description;
+        super(name, description);
     }
 }
