+++ app/src/main/java/com/ss20/se2/monopoly/pojo/models/deeds/Deed.java
@@ -1,22 +1,14 @@
 package com.ss20.se2.monopoly.pojo.models.deeds;
 
-public abstract class Deed {
-    private String name;
+public abstract class Deed extends GameTile {
     private int price;
     private int mortgage;
 
-    public Deed(String name, int price, int mortgage) {
-        this.name = name;
+    Deed(String name, int price, int mortgage) {
+        super(name);
+
         this.price = price;
         this.mortgage = mortgage;
-    }
-
-    public Deed(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return this.name;
     }
 
     public int getPrice() {
